backend/pyproject.toml
[project]
name = "Medkhan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.6",
    "email-validator>=2.2.0",
    "loguru>=0.7.3",
    "yarl>=1.18.3",
    "ruff>=0.9.4",
    "mypy>=1.15.0",
    "sqlalchemy[asyncio]>=2.0.37",
    "alembic>=1.14.1",
    "pytest>=8.3.4",
    "pytest-env>=1.1.5",
    "pytest-cov>=6.1.1",
    "httpx==0.28.1",
    "factory-boy>=3.3.3",
    "redis>=5.2.1",
    "fakeredis>=2.28.1",
    "sqladmin>=0.21.0",
    "python-multipart>=0.0.20",
    "pillow>=11.3.0",
    "aiomysql>=0.2.0",
    "cryptography>=46.0.1",
]

[tool.pyright]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"

include = ["."]
exclude = [
    ".venv",
    "venv",
]

reportMissingImports = "error"
reportUnusedImport = "error"
reportUnusedVariable = "error"

reportUnusedFunction = "warning"
reportUnusedClass = "warning"


[tool.ruff]
line-length = 79
target-version = "py312"

exclude = []

[tool.ruff.lint]
# Здесь выберите только те правила, которые вам нужны, без дублирования.
# `E` и `F` — это стандарт.
# `I` — для сортировки импортов.
# `FA` — для FastAPI.
# `B` — для поиска багов.
# `S` — для безопасности.
# `Q` — для цитат.
# Добавьте, если нужно:
# `TID252` — для запрета относительных импортов
# `TID` - Tidy
# `UP` - Upgrade
select = ["E", "F", "I", "B", "S", "Q"]

fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
